requirements.txt
pyttsx3==2.90
speechrecognition==3.8.1
openai==0.27.0
python-dotenv==0.19.2